requirements.txt
# sudo apt-get install python-dev
--index-url https://pypi.python.org/simple/

pyyaml
mock
